+++ src/main/java/com/github/mafelp/utils/Settings.java
@@ -35,7 +35,7 @@ public class Settings {
     /**
      * version number of the plugin - displayed to users
      */
-    public static final String version = "v0.6.4-beta";
+    public static final String version = "v0.7.0-beta";
 
     /**
      * enables more information being displayed while executing events